requirements.txt
streamlit==1.37.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
